Worked more on 1. I think I'm close to figuring out the basic structure of the Repository, from there I should be able to hammer out the implementation details of that.
My current solution is not entirely decoupled, however, the only coupled parts would be the REST Implementation and SQLite implementation, both of which are fairly set in stone for Android.

Additionally, this coupling should not introduce problems outside of these 2-3 classes for each repository. (Sidenote: Separate repositories with separate implementations for each does seem slightly repetitive)